split bucket init stuff into separate file
